refactor(simulator): extract process scheduler into simulator/process.rs (Step 18)
Move the process-control structs (JoinWaiter, SuspendedProc, AwaitWaiter,
ProcCont + its Drop/impl, ProcessContext) and the scheduler methods
(snapshot/take/restore_process_context, run_scheduled_process[_inner],
run_process_payload, run_process_stmts, proc_await, proc_suspend, proc_resume)
into a dedicated simulator/process.rs submodule. mod.rs shrinks by ~2350 lines.
Structs/fields/methods made pub(crate); sim_debug_enabled replaces the
textual sim_dbg_eprintln! macro (which is not visible across module files).
1935 tests pass.
